[REFACOTR] #63 : 비대한 파일 분리
